requirements.txt
# Core Django and Multi-tenancy
Django==4.2.7
django-tenants==3.6.1

# API and CORS
djangorestframework==3.14.0
django-cors-headers==4.3.1

# Database
psycopg2-binary==2.9.9
dj-database-url==2.1.0

# Caching and Sessions
django-redis==5.4.0

# Media handling
Pillow==10.1.0

# Background tasks
celery==5.3.4
redis==5.0.1

# Configuration
python-decouple==3.8

# Storage (AWS S3)
django-storages==1.14.2
boto3==1.34.0

# AI Integration
openai==1.3.8

# Development and utilities
django-extensions==3.2.3

# Production
gunicorn==21.2.0
whitenoise==6.6.0

# Additional packages for future features
requests==2.31.0
python-slugify==8.0.1
django-crispy-forms==2.1
crispy-bootstrap4==2022.1

# Testing
pytest==7.4.3
pytest-django==4.7.0
factory-boy==3.3.0

# Development tools
black==23.11.0
flake8==6.1.0
isort==5.12.0

# Security
django-ratelimit==4.1.0